requirements.txt
Flask~=2.2.2
Pillow~=9.4.0
selenium~=4.8.0
tqdm~=4.64.1
python-dotenv~=0.21.1
numpy~=1.24.1
requests~=2.28.2